LPS-90507 Move general stuff to general display context
