+++ modules/apps/blogs/blogs-web/src/main/java/com/liferay/blogs/web/internal/display/context/BlogEntriesDisplayContext.java
@@ -15,9 +15,11 @@
 package com.liferay.blogs.web.internal.display.context;
 
 import com.liferay.asset.kernel.model.AssetEntry;
+import com.liferay.blogs.constants.BlogsPortletKeys;
 import com.liferay.blogs.model.BlogsEntry;
 import com.liferay.blogs.service.BlogsEntryLocalServiceUtil;
 import com.liferay.blogs.service.BlogsEntryServiceUtil;
+import com.liferay.blogs.web.internal.security.permission.resource.BlogsEntryPermission;
 import com.liferay.blogs.web.internal.util.BlogsUtil;
 import com.liferay.portal.kernel.dao.search.SearchContainer;
 import com.liferay.portal.kernel.dao.search.SearchContainerResults;
@@ -25,6 +27,8 @@ import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.portlet.LiferayPortletRequest;
+import com.liferay.portal.kernel.portlet.PortalPreferences;
+import com.liferay.portal.kernel.portlet.PortletPreferencesFactoryUtil;
 import com.liferay.portal.kernel.search.Document;
 import com.liferay.portal.kernel.search.Field;
 import com.liferay.portal.kernel.search.Hits;
@@ -33,6 +37,7 @@ import com.liferay.portal.kernel.search.IndexerRegistryUtil;
 import com.liferay.portal.kernel.search.SearchContext;
 import com.liferay.portal.kernel.search.SearchContextFactory;
 import com.liferay.portal.kernel.search.Sort;
+import com.liferay.portal.kernel.security.permission.ActionKeys;
 import com.liferay.portal.kernel.security.permission.PermissionChecker;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.GetterUtil;
@@ -57,7 +62,49 @@ public class BlogEntriesDisplayContext {
 
 		_liferayPortletRequest = liferayPortletRequest;
 
+		_portalPreferences = PortletPreferencesFactoryUtil.getPortalPreferences(
+			liferayPortletRequest);
+
 		_request = _liferayPortletRequest.getHttpServletRequest();
+	}
+
+	public List<String> getAvailableActionDropdownItems(BlogsEntry blogsEntry)
+		throws PortalException {
+
+		List<String> availableActionDropdownItems = new ArrayList<>();
+
+		ThemeDisplay themeDisplay = (ThemeDisplay)_request.getAttribute(
+			WebKeys.THEME_DISPLAY);
+
+		PermissionChecker permissionChecker =
+			themeDisplay.getPermissionChecker();
+
+		if (BlogsEntryPermission.contains(
+			permissionChecker, blogsEntry, ActionKeys.DELETE)) {
+
+			availableActionDropdownItems.add("deleteEntries");
+		}
+
+		return availableActionDropdownItems;
+	}
+
+	public String getDisplayStyle() {
+		String displayStyle = ParamUtil.getString(_request, "displayStyle");
+
+		if (Validator.isNull(displayStyle)) {
+			displayStyle = _portalPreferences.getValue(
+				BlogsPortletKeys.BLOGS_ADMIN, "entries-display-style", "icon");
+		}
+		else {
+			_portalPreferences.setValue(
+				BlogsPortletKeys.BLOGS_ADMIN, "entries-display-style",
+				displayStyle);
+
+			_request.setAttribute(
+				WebKeys.SINGLE_PAGE_APPLICATION_CLEAR_CACHE, Boolean.TRUE);
+		}
+
+		return displayStyle;
 	}
 
 	public void populateResults(SearchContainer searchContainer)
@@ -227,6 +274,7 @@ public class BlogEntriesDisplayContext {
 		BlogEntriesDisplayContext.class);
 
 	private final LiferayPortletRequest _liferayPortletRequest;
+	private final PortalPreferences _portalPreferences;
 	private final HttpServletRequest _request;
 	private Integer _status;
